refactor(dashboard): extract shared helpers to context.go [Phase 1.7]
Move sessionData, formatBytes, relativeTime, and absInt64 from
overview.go and buckets.go into dedicated context.go file. All
dashboard handlers now import shared helpers from one location.
